homework is a little bit redone
