+++ src/main/java/Week1Task3.java
@@ -5,85 +5,89 @@ import java.io.InputStreamReader;
 import java.io.PrintWriter;
 import java.util.*;
 
-
-// D. Правильная скобочная последовательность
+// C. Стек с защитой от ошибок
 public final class Main {
     private Main() {
         // Should not be instantiated
     }
 
     public static class Stack {
-        private char[] data;
+        private int[] data;
         private int last = 0;
 
         public Stack() {
-            this.data = new char[1024];
+            this.data = new int[1024];
         }
 
         private void doubleArray() {
             data = Arrays.copyOf(data, data.length * 2);
         }
 
-        public void push(char c) {
+        private void push(int n) {
             if (data.length <= last) {
                 doubleArray();
             }
-            data[last] = c;
+            data[last] = n;
             last++;
+            System.out.println("ok");
         }
 
-        public int pop() {
+        private void pop() {
             if (last == 0) {
-                return 0;
+                System.out.println("error");
+                return;
             }
-            last--;
+            System.out.println(data[--last]);
             data[last] = 0;
-            return data[last];
         }
 
-        public int back() {
+        private void back() {
             if (last == 0) {
-                return 0;
+                System.out.println("error");
+                return;
             }
-            return data[last - 1];
+            System.out.println(data[last - 1]);
         }
 
-        public int size() {
-            return last;
+        private void size() {
+            System.out.println(last);
         }
 
-        public void clear() {
-            Arrays.fill(data, (char) 0);
+        private void clear() {
+            Arrays.fill(data, 0);
             last = 0;
+            System.out.println("ok");
+        }
+
+        public void exec(String command) {
+            switch (command) {
+                case "pop":
+                    pop();
+                    break;
+                case "back":
+                    back();
+                    break;
+                case "size":
+                    size();
+                    break;
+                case "clear":
+                    clear();
+            }
         }
     }
 
     private static void solve(final FastScanner in, final PrintWriter out) {
         Stack stack = new Stack();
         String s = in.next();
-        char c;
-        for (int i = 0; i < s.length(); i++) {
-            c = s.charAt(i);
-            if (c == '(' || c == '[' || c == '{') {
-                stack.push(c);
+        while (!s.equals("exit")) {
+            if (s.equals("push")) {
+                stack.push(in.nextInt());
             } else {
-                if (stack.size() == 0) {
-                    System.out.println("no");
-                    return;
-                }
-                if ((c == ']' && stack.back() == '[') || (c == ')' && stack.back() == '(') || (c == '}' && stack.back() == '{')) {
-                    stack.pop();
-                } else {
-                    System.out.println("no");
-                    return;
-                }
+                stack.exec(s);
             }
+            s = in.next();
         }
-        if (stack.size() == 0) {
-            System.out.println("yes");
-            return;
-        }
-        System.out.println("no");
+        out.println("bye");
     }
 
     private static class FastScanner {
